Реализация панели админа и одобрения изменения роли на кондитера

+++ src/main/java/ru/kustikov/cakes/producttype/ProductType.java
@@ -16,5 +16,5 @@ public class ProductType {
 
     private boolean canCount;
 
-    private ArrayList<Long> fillings;
+    private ArrayList<Integer> fillings;
 }
